dtDAL: Created a new Bit Mask data type. dtDAL: Created a Bit Mask Actor Property and Named Parameter. dtQt: Created Bit Mask UI Controls. dtQt: Replaced all property control push button widgets with tool buttons.

+++ delta3d/src/dtQt/dynamicactorcontrol.cpp
@@ -165,7 +165,7 @@ namespace dtQt
 
       mTemporaryEditControl->addItem("<None>");
 
-      for (unsigned int i = 0; i < sortedNames.size(); ++i)
+      for (unsigned int i = 0; i < (unsigned int)sortedNames.size(); ++i)
       {
          QString name = sortedNames[i];
